Fixed filenames

+++ update_node.sh
@@ -89,11 +89,18 @@ sleep 10
 
 clear
 
+if ! systemctl status bulwarkd | grep -q "active (running)"; then
+  echo "ERROR: Failed to start bulwarkd. Please contact support."
+  exit
+fi
+
+echo "Installing Bulwark Autoupdater..."
 rm -f /usr/local/bin/bulwarkupdate
 curl -o /usr/local/bin/bulwarkupdate https://raw.githubusercontent.com/bulwark-crypto/Bulwark-MN-Install/master/bulwarkupdate
+chmod a+x /usr/local/bin/bulwarkupdate
 
 if [ ! -f /etc/systemd/system/bulwarkupdate.service ]; then
-cat > /etc/systemd/system/bulwarkd.service << EOL
+cat > /etc/systemd/system/bulwarkupdate.service << EOL
 [Unit]
 Description=Bulwarks's Masternode Autoupdater
 After=network-online.target
@@ -106,7 +113,7 @@ EOL
 fi
 
 if [ ! -f /etc/systemd/system/bulwarkupdate.timer ]; then
-cat > /etc/systemd/system/bulwarkd.timer << EOL
+cat > /etc/systemd/system/bulwarkupdate.timer << EOL
 [Unit]
 Description=Bulwarks's Masternode Autoupdater Timer
 
@@ -121,11 +128,6 @@ fi
 
 systemctl enable bulwarkupdate.timer
 systemctl start bulwarkupdate.timer
-
-if ! systemctl status bulwarkd | grep -q "active (running)"; then
-  echo "ERROR: Failed to start bulwarkd. Please contact support."
-  exit
-fi
 
 echo "Waiting for wallet to load..."
 until su -c "bulwark-cli getinfo 2>/dev/null | grep -q \"version\"" "$USER"; do
